% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power_simulation_functions.R
\name{assign_gRNA}
\alias{assign_gRNA}
\title{Assign gRNAs to Cells Based on Frequency}
\usage{
assign_gRNA(cell_df, gRNA_count_data, missing_prob = 0.2)
}
\arguments{
\item{cell_df}{A data frame representing the cells to which gRNAs will be assigned.}

\item{gRNA_count_data}{gRNA_count_data data frame, including "ID", "target_gene_symbol", and "count" columns.}

\item{missing_prob}{Probability of a cell having missing gRNA information (default: 0.2).}
}
\value{
A modified data frame with gRNA and gene assignments, and a column indicating gRNA detection status.
}
\description{
Assigns gRNAs to cells from a provided list of gRNAs with associated genes,
allowing some cells to have missing gRNA information.
}
\examples{
assigned_df <- assign_gRNA(cell_df, gRNA_count_data = gRNA_count_data, missing_prob = 0.2)
}
