% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power_simulation_functions.R
\name{generate_celltype_df_with_variability}
\alias{generate_celltype_df_with_variability}
\title{Generate Cell Type Data with Variability}
\usage{
generate_celltype_df_with_variability(
  n_cells,
  seurat_obj_metadata = seurat_obj.metadata,
  n_samples,
  sample_col = "orig.ident",
  celltype_col = "NamedClusters",
  enforce_celltype = "Dopamine neurons",
  min_cells = 50
)
}
\arguments{
\item{n_cells}{Total number of cells to generate.}

\item{seurat_obj_metadata}{Metadata from a Seurat object.}

\item{n_samples}{Number of samples to generate.}

\item{sample_col}{Column name in the Seurat metadata that contains sample identifiers (default: "orig.ident").}

\item{celltype_col}{Column name in the Seurat metadata that contains cell type labels (default: "NamedClusters").}

\item{enforce_celltype}{Cell type to enforce with a minimum number of cells (default: "Dopamine neurons").}

\item{min_cells}{Minimum number of cells of the enforced cell type to include in each sample (default: 50).}
}
\value{
A data frame with generated cell types for the specified number of samples.
}
\description{
Generates synthetic cell type data for a specified number of samples, introducing variability in cell type frequencies.
Each sample is created based on the frequencies of cell types in the provided Seurat object metadata.
}
\examples{
cell_df <- generate_celltype_df_with_variability(n_cells = 5000, seurat_obj_metadata = seurat_obj.metadata, n_samples = 5)
}
