% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{gRNA_counts}
\alias{gRNA_counts}
\title{Example gRNA Counts Data}
\format{
A data frame with X rows and Y variables:
\describe{
  \item{target_gene_symbol}{A unique identifier for each gene}
  \item{count}{The count of each gRNA in the sample.}
  \item{X}{Any additional variables relevant to your dataset.}
  \item{ID}{A unique identifier for each gRNA.}
}
}
\source{
Simulated data, or provide a real data source.
}
\usage{
gRNA_counts
}
\description{
This dataset contains example gRNA counts used in the analysis.
Each row corresponds to a gRNA, and each column corresponds to counts.
}
\keyword{datasets}
