% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power_simulation_functions.R
\name{modify_gene_effect}
\alias{modify_gene_effect}
\title{Modify Gene Effect in a Data Frame}
\usage{
modify_gene_effect(
  cell_df,
  nGenes = 5,
  fraction_grnas = 0.8,
  y = 1.5,
  celltype = "Dopamine neurons",
  verbose = TRUE
)
}
\arguments{
\item{cell_df}{A data frame containing information about cells, including "CellType", "gRNA", and "Gene" columns.}

\item{nGenes}{Number of unique genes to modify.}

\item{fraction_grnas}{Fraction of gRNAs for each selected gene to modify.}

\item{y}{The fold change in the number of cells of the specified cell type.}

\item{celltype}{The target cell type to modify (default: "Dopamine neurons").}

\item{verbose}{Logical; if \code{TRUE}, detailed information will be printed.}
}
\value{
A modified data frame where the specified cell type's frequency is adjusted for selected genes.
}
\description{
Modifies the effect of selected genes by adjusting cell type frequencies in a given data frame.
This function simulates the effect of a gene on a specific cell type, altering the frequency of
the cell type in the dataset.
}
\examples{
modified_df <- modify_gene_effect(cell_df, nGenes = 5, fraction_grnas = 0.8, y = 1.5)
}
