% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power_simulation_functions.R
\name{plot_power_simulation}
\alias{plot_power_simulation}
\title{Plot Sensitivity vs. Fraction Cell Proportion Change from Simulation Results}
\usage{
plot_power_simulation(results, palette = get_custom_palette())
}
\arguments{
\item{results}{A data frame containing simulation results. The data frame must have the following columns:
\itemize{
  \item \code{fraction_cell_proportion_change}: Numeric values representing the change in the fraction of cell proportions.
  \item \code{sensitivity}: Numeric values representing the sensitivity of the simulation.
  \item \code{n_samples}: Integer values representing the number of samples used in the simulation.
  \item \code{sd_sensitivity}: Numeric values representing the standard deviation of sensitivity.
  \item \code{nGenes}: Numeric values or factors representing the number of genes in the simulation.
  \item \code{fraction_grnas}: Numeric values or factors representing the fraction of gRNAs in the simulation.
}}

\item{palette}{A vector of colors to be used in the plot. Default is \code{get_custom_palette()}.}
}
\value{
A ggplot object showing the sensitivity vs. fraction cell proportion change plot, faceted by the number of genes and fraction of gRNAs.
}
\description{
This function generates a plot visualizing the relationship between sensitivity and fraction cell proportion change for a simulation study.
The plot shows how sensitivity changes across different sample sizes, while allowing the user to customize the color palette.
}
\details{
This function creates a line and scatter plot using \code{ggplot2} to visualize the sensitivity of the simulation as a function of the fraction of cell proportion change.
The plot is grouped by the number of samples and faceted by the number of genes and fraction of gRNAs. The point size corresponds to the standard deviation of the sensitivity.
}
\examples{
\dontrun{
  # Assuming `simulation_results` is a data frame containing the required columns:
  plot_power_simulation(simulation_results)
}

}
