% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power_simulation_functions.R
\name{power_simulation}
\alias{power_simulation}
\title{Run Power Simulation}
\usage{
power_simulation(
  n_samples,
  n_cells,
  nGenes_vals,
  fraction_grnas_vals,
  fraction_cell_proportion_change_vals,
  n_sims = 10,
  significance_level = 0.05,
  seurat_obj_metadata = seurat_obj.metadata,
  gRNA_count_data
)
}
\arguments{
\item{n_samples}{Number of samples to simulate.}

\item{n_cells}{Number of cells per sample.}

\item{nGenes_vals}{Vector of gene counts to modify.}

\item{fraction_grnas_vals}{Vector of fractions of gRNAs to modify in each gene.}

\item{fraction_cell_proportion_change_vals}{Vector of effect sizes (cell type proportion changes).}

\item{n_sims}{Number of simulations to run.}

\item{significance_level}{The significance level for hypothesis testing.}

\item{seurat_obj_metadata}{Metadata from a Seurat object}

\item{gRNA_count_data}{Path to gRNA data}
}
\value{
A data frame of sensitivity and specificity across simulations.
}
\description{
This function runs power simulations for cell-based experiments with specified parameters.
}
