% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel_setup.R
\name{setup_parallel}
\alias{setup_parallel}
\title{Setup Parallelization}
\usage{
setup_parallel(n_cores = NULL)
}
\arguments{
\item{n_cores}{Number of cores to use. If NULL, uses detectCores() - 1.}
}
\value{
The cluster object created for parallel processing.
}
\description{
Sets up parallel processing with the specified number of cores.
}
