% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel_setup.R
\name{stop_parallel}
\alias{stop_parallel}
\title{Stop Parallelization}
\usage{
stop_parallel(cl)
}
\arguments{
\item{cl}{The cluster object to stop.}
}
\description{
Stops the parallel backend by shutting down the cluster.
}
